requirements.txt
langchain
langchain-community
langchain-ollama
langgraph
langchain-tavily
python-dotenv
matplotlib
yfinance